# docs/swagger.yaml
basePath: /
definitions:
...
  fiber.Map:
    additionalProperties: true
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Appointment:
    properties:
      createdAt:
        type: string
      customer:
        $ref: '#/definitions/models.User'
      customer_id:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_recurring:
        type: boolean
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      recur_pattern:
        allOf:
        - $ref: '#/definitions/models.Recurrence'
        description: E.g., "weekly", "monthly", "daily"
      recur_pattern_id:
        type: integer
      service:
        $ref: '#/definitions/models.Service'
      service_id:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/models.AppointmentStatus'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.AppointmentStatus:
    enum:
    - pending
    - confirmed
    - canceled
    - completed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusCanceled
    - StatusCompleted
...
  models.DayOfWeek:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-varnames:
    - Sunday
    - Monday
    - Tuesday
    - Wednesday
    - Thursday
    - Friday
    - Saturday
  models.Permission:
    properties:
      action:
        description: e.g., "create", "read", "update", "delete"
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        description: e.g., "appointments", "users", etc.
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
...
  models.Recurrence:
    properties:
      appointment_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_after:
        description: Number of occurrences
        type: integer
      frequency:
        description: '"daily", "weekly", "monthly"'
        type: string
      id:
        type: integer
      next_run:
        type: string
      updatedAt:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.Service:
    properties:
      buffer_time:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: Time between appointments
      cost:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      discount:
        description: Discount percentage
        type: number
      discounted_price:
        type: number
      duration:
        $ref: '#/definitions/time.Duration'
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      created_at:
        type: string
      customer_appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      name:
        type: string
      otp:
        type: string
      otp_expires_at:
        type: string
      password:
        type: string
      provided_services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      working_hours:
        items:
          $ref: '#/definitions/models.WorkingHours'
        type: array
    type: object
...
  models.WorkingHours:
    properties:
      break_end:
        description: Optional break end time
        type: string
      break_start:
        description: Optional break start time
        type: string
      createdAt:
        type: string
      day_of_week:
        $ref: '#/definitions/models.DayOfWeek'
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_time:
        description: Format "HH:MM" in 24h
        type: string
      id:
        type: integer
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      start_time:
        description: Format "HH:MM" in 24h
        type: string
      updatedAt:
        type: string
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
  utils.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the Appointment App.
  title: Appointment App API
  version: "1.0"
paths:
  /appointments:
    get:
      consumes:
      - application/json
      description: Retrieve upcoming appointments or appointment history for the authenticated
        user based on status query parameter
      parameters:
      - description: Status filter
        enum:
        - upcoming
        - history
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of appointments
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
        "400":
          description: Bad request - invalid or missing status parameter
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid user token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: No appointments found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get upcoming appointments or appointment history
      tags:
      - appointments
    post:
      consumes:
      - application/json
      description: Create a new appointment with availability checking, working hours
        validation, and email notifications
      parameters:
      - description: Appointment details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "201":
          description: Created appointment
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Service/Customer/Provider not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Time slot conflict or outside working hours
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new appointment
      tags:
      - appointments
  /appointments/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete an appointment from the system
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No content - appointment successfully deleted
        "403":
          description: Forbidden - cannot delete completed or canceled appointment
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an appointment
      tags:
      - appointments
    get:
      consumes:
      - application/json
      description: Retrieve a specific appointment by its ID with associated service,
        provider, and customer details
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Appointment details
          schema:
            $ref: '#/definitions/models.Appointment'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get appointment by ID
      tags:
      - appointments
    patch:
      consumes:
      - application/json
      description: Update appointment details with availability checking and email
        notifications
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated appointment details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: Updated appointment
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Appointment/Service/Customer/Provider not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Time slot conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing appointment
      tags:
      - appointments
  /appointments/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel an existing appointment by changing its status to canceled
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Canceled appointment
          schema:
            $ref: '#/definitions/models.Appointment'
        "403":
          description: Forbidden - cannot cancel completed or already canceled appointment
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel an appointment
      tags:
      - appointments
  /appointments/service/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific service including
        provider and category
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Service details
          schema:
            $ref: '#/definitions/models.Service'
        "404":
          description: Service not found
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get service details by ID
      tags:
      - services
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"